refactor(sc): split connect result handling
